+++ ValheimServersideQoL/Processors/WearNTearProcessor.cs
@@ -1,5 +1,4 @@
-﻿using BepInEx.Logging;
-using static Valheim.ServersideQoL.ModConfig.WearNTearConfig;
+﻿using static Valheim.ServersideQoL.ModConfig.WearNTearConfig;
 
 namespace Valheim.ServersideQoL.Processors;
 
@@ -8,7 +7,7 @@ sealed class WearNTearProcessor : Processor
     protected override bool ProcessCore(ExtendedZDO zdo, IReadOnlyList<Peer> peers)
     {
         UnregisterZdoProcessor = true;
-        if (zdo.PrefabInfo.WearNTear is null || zdo.PrefabInfo.TeleportWorld is not null)
+        if (zdo.PrefabInfo.WearNTear is null)
             return false;
 
         var fields = zdo.Fields<WearNTear>();